requirements.txt
click
pillow
httpx
